refactor(Position): PositionConverter를 삭제하고 해당 역할을 Position 내부로 위임
