+++ src/main/java/chess/domain/position/Position.java
@@ -5,6 +5,11 @@ import java.util.Objects;
 
 public final class Position {
 
+    private static final int COLUMN_INDEX = 0;
+    private static final int ASCII_SMALL_A = 97;
+    private static final int ROW_INDEX = 1;
+    private static final int BOARD_MAX_ROW = 8;
+
     private final Row row;
     private final Column column;
 
@@ -28,6 +33,18 @@ public final class Position {
         Column nextColumn = column.add(unitVector.getColumn());
 
         return new Position(nextRow, nextColumn);
+    }
+
+    public static Position toPosition(final String input) {
+        String value = input.toLowerCase();
+
+        char columnValue = value.charAt(COLUMN_INDEX);
+        int column = Math.abs(columnValue - ASCII_SMALL_A);
+
+        char rowValue = value.charAt(ROW_INDEX);
+        int row = Math.abs(Character.getNumericValue(rowValue) - BOARD_MAX_ROW);
+
+        return new Position(row, column);
     }
 
     public int getRow() {
