+++ src/test/java/chess/domain/PositionTest.java
@@ -1,6 +1,7 @@
 package chess.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import chess.domain.math.UnitVector;
 import chess.domain.position.Position;
@@ -88,5 +89,17 @@ class PositionTest {
         var b = new Position(1, 1);
 
         assertThat(a.equals(b)).isTrue();
+    }
+
+    @Test
+    @DisplayName("사용자의 입력을 2차원 리스트의 좌표값으로 변환해준다.")
+    void toPositionTest_row() {
+        String input = "b2";
+        Position position = Position.toPosition(input);
+
+        assertAll(
+                () -> assertThat(position.getRow()).isEqualTo(6),
+                () -> assertThat(position.getColumn()).isEqualTo(1)
+        );
     }
 }
